+++ Artwork_Management/Views/TabViews/ParentTabView.swift
@@ -47,7 +47,7 @@ struct ParentTabView: View {
                     Spacer(minLength: 0)
                     
                     TabView(selection: $inputTab.selectionTab) {
-                        NewHomeView(itemVM: itemVM, homeVM: homeVM, inputTab: $inputTab)
+                        HomeTabView(itemVM: itemVM, homeVM: homeVM, inputTab: $inputTab)
                             .tag(Tab.home)
 
                         NewItemsView(itemVM: itemVM,  cartVM: cartVM, inputTab: $inputTab)
